refactor: moved ActorPhotos props info to the mocks folder
