+++ reactjs-films-homework/src/pages/ActorProfile/ActorPhotos/__tests__/ActorPhotos.spec.tsx
@@ -5,34 +5,14 @@ import ActorPhotos from '../ActorPhotos';
 
 import { Provider } from 'react-redux';
 import store from '../../../../store/store';
+import ActorPhotosMocks from '../mocks';
 
 describe('ActorPhotos', () => {
-  const mockProps = [
-    {
-      aspect_ratio: 0.6666666666666666,
-      file_path: '/oIciQWr8VwKoR8TmAw1owaiZFyb.jpg',
-      height: 750,
-      iso_639_1: null,
-      vote_average: 5.522,
-      vote_count: 4,
-      width: 500,
-    },
-    {
-      aspect_ratio: 0.6666666666666666,
-      file_path: '/1ahENoyEgKSbRUd0IsydIff7rt1.jpg',
-      height: 3000,
-      iso_639_1: null,
-      vote_average: 5.318,
-      vote_count: 3,
-      width: 2000,
-    },
-  ];
-
   let Fragment: any;
   beforeEach(() => {
     const { asFragment } = render(
       <Provider store={store}>
-        <ActorPhotos photos={mockProps} photosLength={2} />
+        <ActorPhotos photos={ActorPhotosMocks} photosLength={2} />
       </Provider>
     );
     Fragment = asFragment();
