+++ reactjs-films-homework/src/pages/ActorProfile/ActorPhotos/mocks/index.ts
@@ -1,0 +1,1 @@
+export { default } from './ActorPhotosMocks';
